refactor post form components
