+++ apps/cms/src/app/cms/blog/posts/__tests__/MainImageField.spec.tsx
@@ -1,0 +1,19 @@
+import { render, screen } from "@testing-library/react";
+import MainImageField from "@cms/app/cms/blog/posts/MainImageField";
+
+jest.mock("@ui", () => ({
+  Button: ({ children }: any) => <button>{children}</button>,
+  ImagePicker: ({ children }: any) => <div>{children}</div>,
+}));
+
+describe("MainImageField", () => {
+  it("shows select label when empty", () => {
+    render(<MainImageField value="" onChange={() => {}} />);
+    expect(screen.getByText("Select image")).toBeInTheDocument();
+  });
+  it("shows preview when value present", () => {
+    render(<MainImageField value="/test.png" onChange={() => {}} />);
+    expect(screen.getByAltText("Main image")).toHaveAttribute("src", "/test.png");
+    expect(screen.getByText("Change image")).toBeInTheDocument();
+  });
+});
